Migrate BillDIsplay content to dashboard
