+++ src/views/BillDisplay/styles.scss
@@ -23,6 +23,7 @@ div.bill__wrapper {
 
   div.bill__section {
     width: 100%;
+    margin: 20px 10px;
 
     div.bill__list__section {
       margin-left: 2.74117026885%;
@@ -51,7 +52,9 @@ div.bill__wrapper {
         flex-wrap: nowrap;
         flex-direction: column;
         margin-bottom: 15px;
-        box-shadow: -1px -1px 2px rgba(90, 97, 105, 0.25);
+        overflow-y: scroll;
+        height: calc(100vh - 150px);
+        // box-shadow: -1px -1px 2px rgba(90, 97, 105, 0.25);
       }
 
       div.bill__summary {
